+++ src/lang/PropClass.ts
@@ -1,5 +1,6 @@
 import Class from "./Class";
 import Objects from "../types/Objects";
+import { ObjectProps } from "../types/Objects";
 
 export interface IPropClass {
     props: any
@@ -12,7 +13,14 @@ export interface IPropClass {
  */
 abstract class PropsClass extends Class implements IPropClass {
     props: any;
-    constructor(props: any, defaultProps?: any) {
+
+    /**
+     *
+     *
+     * @param props {ObjectsProps}
+     * @param defaultProps {ObjectsProps} it is optional value. It is not exist then checked subclass.defaultProps
+     */
+    protected constructor(props: ObjectProps, defaultProps?: ObjectProps) {
         super();
         let defProps = defaultProps || this.constructor["defaultProps"];
         this.props = defProps ?  Objects.mergeDefaults(defProps, props): props;