+++ src/types/Objects.ts
@@ -1,12 +1,12 @@
 import { has } from "../util/Functions";
 import Type from "../lang/Type";
 import Types from "../util/Types";
-import Assertions from "../util/Assertions";
+import Validations from "../util/Validations";
 
 /**
  * Object interface
  */
-export interface ObjectsProps {
+export interface ObjectProps {
     [key: string]: any
 }
 
@@ -17,7 +17,7 @@ export interface ObjectsProps {
  * ForEach Callback interface
  */
 export interface ForEachCallback {
-    (item: any, key?: string, obj?: ObjectsProps): boolean | void
+    (item: any, key?: string, obj?: ObjectProps): boolean | void
 }
 
 /**
@@ -30,7 +30,7 @@ export interface MapCallback {
     (item: any, key?: string, obj?: Object): any
 }
 /**
- * A class which provides some operations on Object {@see ObjectsProps}
+ * A class which provides some operations on Object {@see ObjectProps}
  * @export
  * @class Objects
  */
@@ -42,7 +42,7 @@ export default class Objects {
      * @param key? {string}
      * @return {boolean}
      */
-    static has<T extends ObjectsProps>(src: T, key?: string): boolean {
+    static has<T extends ObjectProps>(src: T, key?: string): boolean {
         return has(src) && ( has(key) ? src.hasOwnProperty(key): Objects.getLength(src) > 0);
     }
     /**
@@ -50,7 +50,7 @@ export default class Objects {
      * @param src
      * @returns {number}
      */
-    public static getLength<T extends ObjectsProps>(src: T): number {
+    public static getLength<T extends ObjectProps>(src: T): number {
         let i = 0;
         if(!has(src)) return i;
         for(let key in  src) {
@@ -65,7 +65,7 @@ export default class Objects {
      * @param key {any}
      * @return {T}
      */
-    public static remove<T extends ObjectsProps>(src: T, key: any): T {
+    public static remove<T extends ObjectProps>(src: T, key: any): T {
         if(!has(src)) return src;
         delete src[key];
         return src;
@@ -77,7 +77,7 @@ export default class Objects {
      * @param value {any}
      * @return {T}
      */
-    public static removeValue<T extends ObjectsProps>(src: T, value: any): T {
+    public static removeValue<T extends ObjectProps>(src: T, value: any): T {
         if(!has(src)) return src;
         for(let key in  src) {
             if(src.hasOwnProperty(key)) {
@@ -95,7 +95,7 @@ export default class Objects {
      * @param callback
      * @returns {Array<T>}
      */
-    public static map<T extends ObjectsProps, M extends ObjectsProps>(obj: Object, callback: MapCallback ): any[] {
+    public static map<T extends ObjectProps, M extends ObjectProps>(obj: Object, callback: MapCallback ): any[] {
         let result = [];
         if(!obj) return result;
         for(let key in  obj) {
@@ -115,7 +115,7 @@ export default class Objects {
      * @param callback
      * @returns {boolean}
      */
-    public static forEach<T extends ObjectsProps>(obj: T, callback: ForEachCallback): boolean {
+    public static forEach<T extends ObjectProps>(obj: T, callback: ForEachCallback): boolean {
         if(!has(obj)) return true;
         for(let key in  obj) {
             if(obj.hasOwnProperty(key)) {
@@ -132,7 +132,7 @@ export default class Objects {
      * @param src {T}
      * @returns {string[]}
      */
-    public static getKeys<T extends ObjectsProps>(src: T): string[] {
+    public static getKeys<T extends ObjectProps>(src: T): string[] {
         let keys: string[] = [];
         if(!has(src)) return keys;
         for(let key in src) {
@@ -149,10 +149,10 @@ export default class Objects {
      * @param keys? {string[]}
      * @return {T}
      */
-    public static addValue<T extends ObjectsProps>(src: T, key: string, value: any, keys?: string[]): T {
+    public static addValue<T extends ObjectProps>(src: T, key: string, value: any, keys?: string[]): T {
         if(keys && keys.length > 0) {
             let prop = src[key];
-            if(!Assertions.isObject(prop)) prop = src[key] = {};
+            if(!Validations.isObject(prop)) prop = src[key] = {};
             let childKey = keys[0];
             keys.shift();
             Objects.addValue(prop, childKey, value, keys);
@@ -170,7 +170,7 @@ export default class Objects {
      * @param keys? {string[]}
      * @returns {T}
      */
-    public static getValue<T extends ObjectsProps>(src: T, key: string, keys?: string[]): T  {
+    public static getValue<T extends ObjectProps>(src: T, key: string, keys?: string[]): T  {
         let prop = src[key];
         if(prop && keys && keys.length > 0) {
             for(let i = 0 ; i < keys.length; i++) {
@@ -187,7 +187,7 @@ export default class Objects {
      * @param ignoreList? {string[]}
      * @return {T}
      */
-    public static clone<T extends ObjectsProps>(src: T, ignoreList?: string[]): T {
+    public static clone<T extends ObjectProps>(src: T, ignoreList?: string[]): T {
         return Types.getClone(src, ignoreList);
     }
 
@@ -198,13 +198,13 @@ export default class Objects {
      * @param ignoreList? {string[]} specifies get clone of the element by type or not clones just return its.
      * @returns {any}
      */
-    public static merge<S extends ObjectsProps, D extends ObjectsProps>(src: S, dest: D, ignoreList?: string[]): S | D {
+    public static merge<S extends ObjectProps, D extends ObjectProps>(src: S, dest: D, ignoreList?: string[]): S | D {
         if (src == null) return dest;
         if (dest == null) return Objects.clone(src,ignoreList);
-        if(Assertions.isObject(src)) {
+        if(Validations.isObject(src)) {
             for (let key in src) {
                 if (src.hasOwnProperty(key)) {
-                    if (Assertions.isObject(src[key]) && Assertions.isObject(dest[key])) {
+                    if (Validations.isObject(src[key]) && Validations.isObject(dest[key])) {
                         Objects.merge(src[key], dest[key])
                     } else{
                         dest[key] = src[key];
@@ -221,14 +221,14 @@ export default class Objects {
      * @param props {P}
      * @return {any}
      */
-    public static mergeDefaults<D extends ObjectsProps, P extends ObjectsProps>(defaults: D, props: P): D | P {
+    public static mergeDefaults<D extends ObjectProps, P extends ObjectProps>(defaults: D, props: P): D | P {
         if (defaults == null) return props;
         if (props == null) return Objects.clone(defaults,[]);
-        if(Assertions.isObject(defaults)) {
+        if(Validations.isObject(defaults)) {
             for (let key in defaults) {
                 if (defaults.hasOwnProperty(key)) {
                     if(props[key]) {
-                        if(Assertions.isObject(defaults[key]) && Assertions.isObject(props[key])) {
+                        if(Validations.isObject(defaults[key]) && Validations.isObject(props[key])) {
                             props[key] = Objects.mergeDefaults(defaults[key], props[key]);
                         }
                     } else {