+++ src/util/index.ts
@@ -1,5 +1,6 @@
 export * from "./Functions";
 export { default as Types } from "./Types";
+export { default as Validations } from "./Validations";
 export { default as Assertions } from "./Assertions";
 export * from "./Resolver";