+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 * Changed method names and added comments to the **types/Objects** class
 * Changed method names and added comments to the **types/Arrays** class
 * Changed method names and added comments to the **types/Strings** class
+* Changed method names and added comments to the **types/Assertions** class
+* Changed method names and added comments to the **types/Functions** class
+* Changed method names and added comments to the **types/Resolver** class
+* Changed method names and added comments to the **types/Types** class
+* Added **types/Validations** class.
+* Changed method names and added comments to the **types/Strings** class
+* Changed method names and added comments to the **types/Strings** class
 
 #### v1.0.2
 * fixed document link problems.